+++ rio/ChangeLog.md
@@ -3,6 +3,9 @@
 ## 0.1.3.0
 
 * Add `newLogFunc` function to create `LogFunc` records outside of a callback scope
+* Add `logGenericCallStack` to be able to register a log entry `CallStack` from a
+  different thread
+* Add `getLogMinLevel` to gather what is the log min level from `LogOptions`
 * Allow dynamic reloading of `logMinLevel` and `logVerboseFormat` for the `LogOptions` record
 * Add `headMaybe`, `lastMaybe`, `tailMaybe`, `initMaybe`, `maximumMaybe`, `minimumMaybe`,
   `maximumByMaybe`, `minimumByMaybe` functions to `RIO.List` module (issue #82)